cmd/http/api/category/handler/get_category_list_test.go
package handler

import (
	"context"
	"testing"

	category "git.zqbjj.top/pet/services/cmd/http/dto/hertz_gen/category"
	common "git.zqbjj.top/pet/services/cmd/http/dto/hertz_gen/common"
	"github.com/cloudwego/hertz/pkg/app"
	"github.com/cloudwego/hertz/pkg/common/test/assert"
)

func TestGetCategoryListService_Run(t *testing.T) {
	ctx := context.Background()
	c := app.NewContext(1)
	s := NewGetCategoryListService(ctx, c)
	// init req and assert value
	req := &common.PageFilter{}
	resp, err := s.Do(req)
	assert.DeepEqual(t, nil, resp)
	assert.DeepEqual(t, nil, err)
	// todo edit your unit test.
}
